Check config

+++ sidecar/sidecar.py
@@ -7,10 +7,13 @@ import requests
 import re
 import shutil
 import tempfile
+import subprocess
 from urllib3.util.retry import Retry
 from requests.adapters import HTTPAdapter
 
 prog = re.compile('^(?P<file_name>.+)\.part\d+\.?.*$')
+prog2 = re.compile('^(?P<file_name>.+)-cmid-.*$')
+
 sourceFolder = tempfile.mkdtemp()
 transform={}
 
@@ -96,20 +99,21 @@ def watchForChanges(label, targetFolder, url, method, payload, current):
                 print("File in configmap %s %s" % (filename, eventType))
                 if (eventType == "ADDED") or (eventType == "MODIFIED"):
                     if partfiles:
-                        if prog.match(filename):  #if parts\d+ is in file 
-                            writeTextToFile(sourceFolder,filename+cmid , dataMap[filename])
-                        else:
-                            writeTextToFile(targetFolder,filename, dataMap[filename])
+                        #if prog.match(filename):  #if parts\d+ is in file 
+                            writeTextToFile(sourceFolder,filename+"-cmid-"+cmid , dataMap[filename])
+                        #else:
+                        #    writeTextToFile(targetFolder,filename, dataMap[filename])
                     else:
                       writeTextToFile(targetFolder, filename, dataMap[filename])
 
                     
                 else:
                     rmFile(sourceFolder,targetFolder,filename)
-
+            reloadConfig=True
             if partfiles:
-                processFiles(sourceFolder,targetFolder,comment)
-            if url is not None:
+                reloadConfig=processFiles(sourceFolder,targetFolder,comment)
+            
+            if url is not None and reloadConfig:
                        request(url, method, payload)
 
 # $Env:DATA_NAME_ROUTE="alertmanager-route"
@@ -180,9 +184,13 @@ def appendParts(listfiles,frompath,topath,comment):
   for f in listfiles:
     file_name=f
 
-    result = prog.match(file_name)
+    result = prog.match(file_name) # parts
     if result:
       file_name=result.group('file_name')
+    elif prog2.match(file_name): #hole files
+      result2=prog2.match(file_name)
+      file_name=result2.group('file_name')
+
     fromfullpath=os.path.join(frompath, f)
     fromfile = open(fromfullpath, 'r') 
     tofullpath=os.path.join(topath, file_name)
@@ -199,6 +207,34 @@ def indent(text, count_ident=0):
     indent=' ' * count_ident
     return ''.join([indent + l for l in text.splitlines(True)])
 
+def checkConfig():
+    
+    command = os.getenv('CHECK_CONFIG_COMMAND')
+    if command is None:
+      return True
+
+    print("Start check config")
+    print("Command:" + command)
+    
+    ok_exit_codes = os.getenv('OK_EXIT_CODES')
+    if ok_exit_codes is None:
+      ok_exit_codes='0,127'
+
+    return_code = subprocess.call(command, shell=True)
+    print("Return code:"+str(return_code))
+    config_ok=False
+    for code in ok_exit_codes.split(','):
+        if int(code) == int(return_code):
+          config_ok=True
+
+        if config_ok:
+            print(" Config is ok ")
+            return True
+        else:
+            print("Oups Something wrong ")
+            return False
+
+
 def copyToDest(src,dest):
     src_files = os.listdir(src)
     for file_name in src_files:
@@ -208,17 +244,30 @@ def copyToDest(src,dest):
         print("Created "+file_name + " in " + dest)
 
 def processFiles(sourcepath,destpath,comment):
-  tmppath=tempfile.mkdtemp()
-  listall = [f for f in os.listdir(sourcepath)]
-  listall.sort()
+  tmp=tempfile.gettempdir()
+  #tmppath=tempfile.mkdtemp()
+  tmppath=os.path.join(tmp,"k8s_sidecar-cm-to-file")
+  if os.path.exists(tmppath):
+      shutil.rmtree(tmppath)
+  os.makedirs(tmppath)
+  print("Temp directory for copy after check config is: "+ tmppath)
+  listall = getFiles(sourcepath)
+  #rename alias to file name
   transform=getTransform()
   if transform :
       transformFiles(transform,listall,sourcepath)
-      listall = [f for f in os.listdir(sourcepath)]
-      listall.sort()
+      listall = getFiles(sourcepath)
+  #procces 
   appendParts(listall,sourcepath,tmppath,comment)
-  copyToDest(tmppath,destpath)
-  shutil.rmtree(tmppath)
+  check_config=checkConfig()
+  if check_config:
+     copyToDest(tmppath,destpath)
+  return check_config
+
+def getFiles(path):
+    listall = [f for f in os.listdir(path)]
+    listall.sort()
+    return listall
 
 def main():
     print("Starting config map collector")
